pyproject.toml
[tool.poetry]
name = "wallettest"
version = "0.1.0"
description = "REST API server using **django-rest-framework** with pagination, sorting and filtering for two models:"
authors = ["Anton Bogunkov <anton.bogunkov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
environs = { extras = ["django"], version = "^11.2.1" }
Django = "^5.1.3"
djangorestframework = "^3.15.2"
djangorestframework-jsonapi = "^7.1.0"
django-cors-headers = "^4.6.0"
mysqlclient = "^2.2.6"
django-filter = "^24.3"
drf-spectacular-jsonapi = "^0.5.2"
drf-spectacular = "^0.27.2"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"
factory-boy = "^3.3.1"
coverage = "^7.6.8"
diff-cover = "^9.2.0"
pylint = "^3.3.1"
pylint-django = "^2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
